requirements.txt
click
flake8
ipywidgets
jupyterlab
jupyterlab-vim
h5py
matplotlib
monai
mypy
nibabel
neurite
numpy
pandas
pillow
pydicom
python-gdcm
scipy
scikit-image
scikit-learn
SimpleITK
tensorflow
torch
torchvision
tqdm
voxelmorph
